Updated the Code

+++ src/main/java/com/ecommerce/dto/CartItemDTO.java
@@ -1,5 +1,6 @@
 package com.ecommerce.dto;
 
+import java.math.BigDecimal;
 import lombok.Data;
 
 @Data
@@ -7,11 +8,11 @@ public class CartItemDTO {
     private Long id;
     private Long productId;
     private String productName;
-    private Double unitPrice;
+    private BigDecimal unitPrice;
     private Integer quantity;
     private String imageUrl;
 
-    public CartItemDTO(Long id, Long productId, String productName, Double unitPrice, Integer quantity, String imageUrl) {
+    public CartItemDTO(Long id, Long productId, String productName, BigDecimal unitPrice, Integer quantity, String imageUrl) {
         this.id = id;
         this.productId = productId;
         this.productName = productName;
